Wait for install to complete before proceeding
There was a bug where the script would say that passly did not install successfully even though it did. This happened because simply calling msi exec does not wait for the install to complete before moving on. Changing it so we use Start-Process with the -wait parameter. This makes it so the script waits for the program to be installed before moving on.